Move to own query-tabs component
